+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Nested dialog resets `body` padding
 - Inherit `namespace` value from parent `impulse_form_with`
 
 ## [0.4.0] - 2024-03-05